Put sov tx inspection and short header proof update to common utils
